feat:Create Teste for Generics and update classes

+++ src/main/java/br/com/felipecesar/tests_junit_mockito/repository/TodoRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TodoRepository extends JpaRepository<TodoEntity, Integer> {
+public interface TodoRepository extends JpaRepository<TodoEntity, Long> {
 }
